requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
supabase==2.10.0
pydantic-settings==2.7.1
python-dotenv==1.0.1
tenacity==9.0.0
httpx==0.27.2
PyJWT==2.10.1
cryptography==43.0.3
pytest==8.3.3

